+++ src/ModelToSearchThrough.php
@@ -30,17 +30,38 @@ class ModelToSearchThrough
     protected int $key;
 
     /**
+     * Full-text search.
+     */
+    protected bool $fullText;
+
+    /**
+     * Full-text search options
+     */
+    protected array $fullTextOptions = [];
+
+    /**
+     * Full-text through relation.
+     */
+    protected ?string $fullTextRelation = null;
+
+    /**
      * @param \Illuminate\Database\Eloquent\Builder $builder
      * @param \Illuminate\Support\Collection $columns
      * @param string $orderByColumn
      * @param integer $key
+     * @param bool $fullText
+     * @param array $fullTextOptions
+     * @param string $fullTextRelation
      */
-    public function __construct(Builder $builder, Collection $columns, string $orderByColumn, int $key)
+    public function __construct(Builder $builder, Collection $columns, string $orderByColumn, int $key, bool $fullText = false, array $fullTextOptions = [], string $fullTextRelation = null)
     {
-        $this->builder       = $builder;
-        $this->columns       = $columns;
-        $this->orderByColumn = $orderByColumn;
-        $this->key           = $key;
+        $this->builder          = $builder;
+        $this->columns          = $columns;
+        $this->orderByColumn    = $orderByColumn;
+        $this->key              = $key;
+        $this->fullText         = $fullText;
+        $this->fullTextOptions  = $fullTextOptions;
+        $this->fullTextRelation = $fullTextRelation;
     }
 
     /**
@@ -64,6 +85,28 @@ class ModelToSearchThrough
     public function getFreshBuilder(): Builder
     {
         return clone $this->builder;
+    }
+
+    /**
+     * Get a collection with all columns or relations to search through.
+     *
+     * @return \Illuminate\Support\Collection
+     */
+    public function getColumns(): Collection
+    {
+        return $this->columns;
+    }
+
+    /**
+     * Set a collection with all columns or relations to search through.
+     *
+     * @return $this
+     */
+    public function setColumns(Collection $columns): self
+    {
+        $this->columns = $columns;
+
+        return $this;
     }
 
     /**
@@ -109,7 +152,7 @@ class ModelToSearchThrough
      * @param string $column
      * @return string
      */
-    protected function qualifyColumn(string $column): string
+    public function qualifyColumn(string $column): string
     {
         return $this->getModel()->qualifyColumn($column);
     }
@@ -132,5 +175,79 @@ class ModelToSearchThrough
     public function getQualifiedOrderByColumnName(): string
     {
         return $this->qualifyColumn($this->orderByColumn);
+    }
+
+    /**
+     * Full-text search.
+     *
+     * @return boolean
+     */
+    public function isFullTextSearch(): bool
+    {
+        return $this->fullText;
+    }
+
+    /**
+     * Full-text search options.
+     *
+     * @return array
+     */
+    public function getFullTextOptions(): array
+    {
+        return $this->fullTextOptions;
+    }
+
+    /**
+     * Full-text through relation.
+     *
+     * @return string|null
+     */
+    public function getFullTextRelation(): ?string
+    {
+        return $this->fullTextRelation;
+    }
+
+    /**
+     * Full-text through relation.
+     *
+     * @return $this
+     */
+    public function setFullTextRelation(?string $fullTextRelation = null): self
+    {
+        $this->fullTextRelation = $fullTextRelation;
+
+        return $this;
+    }
+
+    /**
+     * Clone the current instance.
+     *
+     * @return static
+     */
+    public function clone(): static
+    {
+        return new static($this->builder, $this->columns, $this->orderByColumn, $this->key, $this->fullText, $this->fullTextOptions, $this->fullTextRelation);
+    }
+
+    /**
+     * Split the current instance into multiple based on relation search.
+     *
+     * @return \Illuminate\Support\Collection
+     */
+    public function toGroupedCollection(): Collection
+    {
+        if ($this->columns->all() === $this->columns->flatten()->all()) {
+            return Collection::wrap($this);
+        }
+
+        $collection = Collection::make();
+
+        foreach ($this->columns as $relation => $columns) {
+            $collection->push(
+                $this->clone()->setColumns(Collection::wrap($columns))->setFullTextRelation($relation)
+            );
+        }
+
+        return $collection;
     }
 }