+++ tests/Page.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ProtoneMedia\LaravelCrossEloquentSearch\Tests;
+
+use Illuminate\Database\Eloquent\Model;
+
+class Page extends Model
+{
+}