+++ tests/SearchTest.php
@@ -6,7 +6,9 @@ use Illuminate\Contracts\Pagination\LengthAwarePaginator;
 use Illuminate\Pagination\Paginator;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Collection;
+use ProtoneMedia\LaravelCrossEloquentSearch\OrderByRelevanceException;
 use ProtoneMedia\LaravelCrossEloquentSearch\Search;
+use ProtoneMedia\LaravelCrossEloquentSearch\Searcher;
 
 class SearchTest extends TestCase
 {
@@ -25,7 +27,7 @@ class SearchTest extends TestCase
 
         $results = Search::add(Post::class, 'title')
             ->add(Video::class, 'title')
-            ->get('foo');
+            ->search('foo');
 
         $this->assertInstanceOf(Collection::class, $results);
         $this->assertCount(2, $results);
@@ -49,7 +51,7 @@ class SearchTest extends TestCase
 
         $results = Search::add(Post::class, 'title')
             ->add(Video::class, ['title', 'subtitle'])
-            ->get('foo');
+            ->search('foo');
 
         $this->assertCount(3, $results);
 
@@ -98,7 +100,7 @@ class SearchTest extends TestCase
 
         $results = Search::add(Post::class, 'title')
             ->add(Video::class, 'title')
-            ->get('"bar bar"');
+            ->search('"bar bar"');
 
         $this->assertCount(1, $results);
 
@@ -116,7 +118,7 @@ class SearchTest extends TestCase
         $results = Search::add(Post::class, 'title')
             ->add(Video::class, 'title')
             ->dontParseTerm()
-            ->get('bar bar');
+            ->search('bar bar');
 
         $this->assertCount(1, $results);
 
@@ -136,7 +138,7 @@ class SearchTest extends TestCase
             ->add(VideoJson::class, 'title->nl')
             ->beginWithWildcard()
             ->ignoreCase()
-            ->get('FOO');
+            ->search('FOO');
 
         $this->assertCount(2, $results);
     }
@@ -170,7 +172,7 @@ class SearchTest extends TestCase
         $results = Search::new()
             ->add(Post::class)->orderBy('updated_at')
             ->add(Video::class)->orderBy('published_at')
-            ->get();
+            ->search();
 
         $this->assertCount(4, $results);
     }
@@ -184,9 +186,9 @@ class SearchTest extends TestCase
         Video::create(['title' => 'bar']);
 
         $results = Search::new()
-            ->addWhen(true, Post::class, 'title')
-            ->addWhen(false, Video::class, 'title', 'published_at')
-            ->get('foo');
+            ->when(true, fn (Searcher $searcher) => $searcher->add(Post::class, 'title'))
+            ->when(false, fn (Searcher $searcher) => $searcher->add(Video::class, 'title', 'published_at'))
+            ->search('foo');
 
         $this->assertCount(1, $results);
         $this->assertTrue($results->first()->is($postA));
@@ -201,7 +203,7 @@ class SearchTest extends TestCase
         $results = Search::addMany([
             [Video::class, 'title'],
             [Video::class, 'subtitle', 'created_at'],
-        ])->get('foo');
+        ])->search('foo');
 
         $this->assertCount(2, $results);
 
@@ -214,8 +216,8 @@ class SearchTest extends TestCase
     {
         Video::create(['title' => 'foo']);
 
-        $this->assertCount(1, Search::add(Video::class, 'title')->get('fo'));
-        $this->assertCount(0, Search::add(Video::class, 'title')->endWithWildcard(false)->get('fo'));
+        $this->assertCount(1, Search::add(Video::class, 'title')->search('fo'));
+        $this->assertCount(0, Search::add(Video::class, 'title')->endWithWildcard(false)->search('fo'));
     }
 
     /** @test */
@@ -223,8 +225,8 @@ class SearchTest extends TestCase
     {
         Video::create(['title' => 'laravel']);
 
-        $this->assertCount(0, Search::add(Video::class, 'title')->get('larafel'));
-        $this->assertCount(1, Search::add(Video::class, 'title')->soundsLike()->get('larafel'));
+        $this->assertCount(0, Search::add(Video::class, 'title')->search('larafel'));
+        $this->assertCount(1, Search::add(Video::class, 'title')->soundsLike()->search('larafel'));
     }
 
     /** @test */
@@ -235,7 +237,7 @@ class SearchTest extends TestCase
 
         $results = Search::add(Video::class, 'title')
             ->add(Video::class, 'subtitle')
-            ->get('foo');
+            ->search('foo');
 
         $this->assertCount(2, $results);
 
@@ -254,7 +256,7 @@ class SearchTest extends TestCase
         $results = Search::add(Post::class, 'title', 'published_at')
             ->add(Video::class, 'title', 'published_at')
             ->orderByDesc()
-            ->get('foo');
+            ->search('foo');
 
         $this->assertCount(2, $results);
 
@@ -272,7 +274,7 @@ class SearchTest extends TestCase
 
         $results = Search::add(Post::whereNotNull('published_at'), 'title')
             ->add(Video::whereNotNull('published_at'), 'title')
-            ->get('foo');
+            ->search('foo');
 
         $this->assertCount(1, $results);
 
@@ -291,8 +293,8 @@ class SearchTest extends TestCase
             ->add(Video::class, 'title', 'published_at')
             ->orderByDesc();
 
-        $resultsPage1 = $search->paginate(2, 'page', 1)->get('foo');
-        $resultsPage2 = $search->paginate(2, 'page', 2)->get('foo');
+        $resultsPage1 = $search->paginate(2, 'page', 1)->search('foo');
+        $resultsPage2 = $search->paginate(2, 'page', 2)->search('foo');
 
         $this->assertInstanceOf(LengthAwarePaginator::class, $resultsPage1);
         $this->assertInstanceOf(LengthAwarePaginator::class, $resultsPage2);
@@ -319,11 +321,76 @@ class SearchTest extends TestCase
         }
 
         $results = Search::add(Post::with('comments')->withCount('comments'), 'title')
-            ->get('foo');
+            ->search('foo');
 
         $this->assertCount(1, $results);
         $this->assertEquals(10, $results->first()->comments_count);
         $this->assertTrue($results->first()->relationLoaded('comments'));
+    }
+
+    /** @test */
+    public function it_can_search_through_relations()
+    {
+        $videoA = Video::create(['title' => 'foo1']);
+        $videoB = Video::create(['title' => 'bar']);
+
+        $postAA = $videoA->posts()->create(['title' => 'foo2']);
+        $postAB = $videoA->posts()->create(['title' => 'bar']);
+        $postBA = $videoB->posts()->create(['title' => 'far']);
+        $postBB = $videoB->posts()->create(['title' => 'boo']);
+
+        $postAA->comments()->create(['body' => 'comment1']);
+        $postAB->comments()->create(['body' => 'comment2']);
+        $postBA->comments()->create(['body' => 'comment3']);
+        $postBB->comments()->create(['body' => 'comment4']);
+
+        $results = Search::new()
+            ->beginWithWildcard(false)
+            ->endWithWildcard(false)
+            ->add(Video::class, 'posts.comments.body')
+            ->search('comment4');
+
+        $this->assertCount(1, $results);
+        $this->assertTrue($results->first()->is($videoB));
+
+        $results = Search::new()
+            ->beginWithWildcard(false)
+            ->endWithWildcard(false)
+            ->add(Video::class, ['title', 'posts.comments.body'])
+            ->search('foo1 comment4');
+
+        $this->assertCount(2, $results);
+
+        $results = Search::new()
+            ->beginWithWildcard(false)
+            ->endWithWildcard(false)
+            ->add(Video::class, ['title', 'posts.comments.body'])
+            ->add(Post::class, ['title', 'comments.body'])
+            ->search('foo1 foo2 comment4');
+
+        $this->assertCount(4, $results);
+
+        $results = $results->map(fn ($model) => class_basename($model) . $model->getKey());
+
+        $this->assertTrue($results->contains('Video1'));    // because foo1
+        $this->assertTrue($results->contains('Video2'));    // because comment4
+        $this->assertTrue($results->contains('Post1'));     // because foo2
+        $this->assertTrue($results->contains('Post4'));     // because comment4
+    }
+
+    /** @test */
+    public function it_doesnt_add_term_constraints_when_the_search_query_is_empty()
+    {
+        $videoA = Video::create(['title' => 'foo']);
+        $videoB = Video::create(['title' => 'bar']);
+
+        $videoA->posts()->create(['title' => 'far']);
+
+        $results = Search::new()
+            ->add(Video::class, ['title', 'posts.title'])
+            ->search();
+
+        $this->assertCount(2, $results);
     }
 
     /** @test */
@@ -340,7 +407,7 @@ class SearchTest extends TestCase
             ->orderByModel([
                 Comment::class, Post::class, Video::class,
             ])
-            ->get('foo');
+            ->search('foo');
 
         $this->assertInstanceOf(Comment::class, $results->get(0));
         $this->assertInstanceOf(Post::class, $results->get(1));
@@ -355,7 +422,7 @@ class SearchTest extends TestCase
                 Post::class, Video::class, Comment::class,
             ])
             ->orderByDesc()
-            ->get('foo');
+            ->search('foo');
 
         $this->assertInstanceOf(Comment::class, $results->get(0));
         $this->assertInstanceOf(Video::class, $results->get(1));
@@ -367,7 +434,7 @@ class SearchTest extends TestCase
             ->add(Video::class, ['title'])
             ->add(Comment::class, ['body'])
             ->orderByModel(Comment::class)
-            ->get('foo');
+            ->search('foo');
 
         $this->assertInstanceOf(Comment::class, $results->get(0));
     }
@@ -384,7 +451,7 @@ class SearchTest extends TestCase
             ->add(Post::class, 'title', 'published_at')
             ->add(Video::class, 'title', 'published_at')
             ->orderByModel([Video::class, Post::class])
-            ->get('foo');
+            ->search('foo');
 
         $this->assertCount(4, $results);
 
@@ -407,11 +474,32 @@ class SearchTest extends TestCase
             ->beginWithWildcard()
             ->orderByRelevance()
             ->orderByModel([Video::class, Post::class])
-            ->get('Apple iPad');
+            ->search('Apple iPad');
 
         $this->assertCount(4, $results);
         $this->assertTrue($results->first()->is($videoB), $results->toJson());
         $this->assertTrue($results->last()->is($postA), $results->toJson());
+    }
+
+    /** @test */
+    public function it_cant_order_by_relevance_when_searching_through_nested_relationships()
+    {
+        $video = Video::create(['title' => 'foo']);
+        $post  = $video->posts()->create(['title' => 'bar']);
+
+        $search = Search::new()
+            ->beginWithWildcard(false)
+            ->endWithWildcard(false)
+            ->add(Video::class, 'posts.title')
+            ->orderByRelevance();
+
+        try {
+            $search->search('bar');
+        } catch (OrderByRelevanceException $e) {
+            return $this->assertTrue(true);
+        }
+
+        $this->fail('Should have thrown OrderByRelevanceException');
     }
 
     /** @test */
@@ -424,7 +512,7 @@ class SearchTest extends TestCase
             ->add(Video::class, ['title', 'subtitle'])
             ->beginWithWildcard()
             ->orderByRelevance()
-            ->get('Apple iPad');
+            ->search('Apple iPad');
 
         $this->assertCount(2, $results);
         $this->assertTrue($results->first()->is($videoB));
@@ -439,7 +527,7 @@ class SearchTest extends TestCase
         $results = Search::new()
             ->add(Video::class)
             ->orderByRelevance()
-            ->get();
+            ->search();
 
         $this->assertCount(2, $results);
     }
@@ -451,7 +539,7 @@ class SearchTest extends TestCase
             ->add(Video::class, 'title', 'published_at')
             ->orderByDesc();
 
-        $results = $search->paginate()->get('foo');
+        $results = $search->paginate()->search('foo');
 
         $this->assertInstanceOf(LengthAwarePaginator::class, $results);
     }
@@ -464,7 +552,7 @@ class SearchTest extends TestCase
             ->add(Video::class, 'title', 'published_at')
             ->orderByDesc();
 
-        $results = $search->simplePaginate()->get('foo');
+        $results = $search->simplePaginate()->search('foo');
 
         $this->assertInstanceOf(Paginator::class, $results);
     }
@@ -482,8 +570,8 @@ class SearchTest extends TestCase
             ->add(Video::class, 'title', 'published_at')
             ->orderByDesc();
 
-        $resultsPage1 = $search->simplePaginate(2, 'page', 1)->get('foo');
-        $resultsPage2 = $search->simplePaginate(2, 'page', 2)->get('foo');
+        $resultsPage1 = $search->simplePaginate(2, 'page', 1)->search('foo');
+        $resultsPage2 = $search->simplePaginate(2, 'page', 2)->search('foo');
 
         $this->assertInstanceOf(Paginator::class, $resultsPage1);
         $this->assertInstanceOf(Paginator::class, $resultsPage2);
@@ -495,5 +583,139 @@ class SearchTest extends TestCase
 
         $this->assertTrue($resultsPage2->first()->is($postA));
         $this->assertTrue($resultsPage2->last()->is($videoA));
+    }
+
+    /** @test */
+    public function it_includes_a_model_identifier_to_search_results()
+    {
+        Post::create(['title' => 'bar']);
+        Video::create(['title' => 'baz']);
+
+        $search = Search::new()
+            ->add(Post::class, 'title', 'title')
+            ->add(Video::class, 'title', 'title')
+            ->includeModelType()
+            ->paginate()
+            ->search('ba');
+
+        $this->assertEquals($search->toArray()['data'][0]['type'], class_basename(Post::class));
+        $this->assertEquals($search->toArray()['data'][1]['type'], class_basename(Video::class));
+    }
+
+    /** @test */
+    public function it_includes_a_custom_model_identifier_to_search_results()
+    {
+        Post::create(['title' => 'bar']);
+        Video::create(['title' => 'baz']);
+
+        $search = Search::new()
+            ->add(VideoJson::class, 'title', 'title')
+            ->includeModelType()
+            ->paginate()
+            ->search('ba');
+
+        $this->assertEquals($search->toArray()['data'][0]['type'], 'awesome_video');
+    }
+
+    /** @test */
+    public function it_supports_full_text_search()
+    {
+        $postA = Post::create(['title' => 'Laravel Framework']);
+        $postB = Post::create(['title' => 'Tailwind Framework']);
+
+        $blogA = Blog::create(['title' => 'Laravel Framework', 'subtitle' => 'PHP', 'body' => 'Ad nostrud adipisicing deserunt labore reprehenderit ']);
+        $blogB = Blog::create(['title' => 'Tailwind Framework', 'subtitle' => 'CSS', 'body' => 'aute do commodo ea magna dolor cupidatat ullamco commodo.']);
+
+        $pageA = Page::create(['title' => 'Laravel Framework', 'subtitle' => 'PHP', 'body' => 'Ad nostrud adipisicing deserunt labore reprehenderit ']);
+        $pageB = Page::create(['title' => 'Tailwind Framework', 'subtitle' => 'CSS', 'body' => 'aute do commodo ea magna dolor cupidatat ullamco commodo.']);
+
+        $results = Search::new()
+            ->beginWithWildcard()
+            ->add(Post::class, 'title')
+            ->addFullText(Blog::class, ['title', 'subtitle', 'body'], ['mode' => 'boolean'])
+            ->addFullText(Page::class, ['title', 'subtitle', 'body'], ['mode' => 'boolean'])
+            ->search('framework -css');
+
+        $this->assertCount(4, $results);
+
+        $this->assertTrue($results->contains($postA));
+        $this->assertTrue($results->contains($postB));
+        $this->assertTrue($results->contains($blogA));
+        $this->assertTrue($results->contains($pageA));
+    }
+
+    /** @test */
+    public function it_supports_full_text_search_on_relations()
+    {
+        $videoA = Video::create(['title' => 'Page A']);
+        $videoB = Video::create(['title' => 'Page B']);
+        $videoC = Video::create(['title' => 'Page C']);
+        $videoD = Video::create(['title' => 'Page D']);
+
+        $videoA->blogs()->create(['title' => 'Laravel Framework', 'subtitle' => 'PHP', 'body' => 'Ad nostrud adipisicing deserunt labore reprehenderit ']);
+        $videoB->blogs()->create(['title' => 'Tailwind Framework', 'subtitle' => 'CSS', 'body' => 'aute do commodo ea magna dolor cupidatat ullamco commodo.']);
+        $videoC->pages()->create(['title' => 'Laravel Framework', 'subtitle' => 'PHP', 'body' => 'Ad nostrud adipisicing deserunt labore reprehenderit ']);
+        $videoD->pages()->create(['title' => 'Tailwind Framework', 'subtitle' => 'CSS', 'body' => 'aute do commodo ea magna dolor cupidatat ullamco commodo.']);
+
+        $results = Search::new()
+            ->beginWithWildcard()
+            ->addFullText(Video::class, [
+                'blogs' => ['title', 'subtitle', 'body'],
+                'pages' => ['title', 'subtitle', 'body'],
+            ], )
+            ->search('framework -css');
+
+        $this->assertCount(2, $results);
+
+        $this->assertTrue($results->contains($videoA));
+        $this->assertTrue($results->contains($videoC));
+    }
+
+    /** @test */
+    public function it_returns_data_consistently()
+    {
+        Carbon::setTestNow(now());
+        $postA = Post::create(['title' => 'Laravel Framework']);
+
+        Carbon::setTestNow(now()->addSecond());
+        $postB = Post::create(['title' => 'Tailwind Framework']);
+
+        $this->assertEquals(2, Post::all()->count());
+        $this->assertEquals(0, Blog::all()->count());
+
+        $resultA = Search::addMany([
+            [Post::query(), 'title'],
+        ])->search('');
+
+        $resultB = Search::addMany([
+            [Post::query(), 'title'],
+            [Blog::query(), 'title'],
+        ])->search('');
+
+        $this->assertCount(2, $resultA);
+        $this->assertCount(2, $resultB);
+
+        $this->assertTrue($resultA->first()->is($postA));
+        $this->assertTrue($resultB->first()->is($postA));
+    }
+
+    /** @test */
+    public function it_can_conditionally_apply_ordering()
+    {
+        Carbon::setTestNow(now());
+        $postA = Post::create(['title' => 'foo']);
+
+        Carbon::setTestNow(now()->subDay());
+        $postB = Post::create(['title' => 'foo2']);
+
+        $results = Search::add(Post::class, 'title')
+            ->when(true, fn (Searcher $searcher) => $searcher->orderByDesc())
+            ->search('foo');
+
+        $this->assertInstanceOf(Collection::class, $results);
+        $this->assertCount(2, $results);
+
+        $this->assertTrue($results->first()->is($postA));
+        $this->assertTrue($results->last()->is($postB));
     }
 }